chore: mobile view of success modal
