+++ packages/cfd/src/Components/migration-success-modal/index.tsx
@@ -1,0 +1,3 @@
+import MigrationSuccessModal from './migration-success-modal';
+
+export default MigrationSuccessModal;
